// src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        Status: {
// ...
            account?: components["schemas"]["Account"];
// ...
        };
// ...
        VisibilityScope: "public" | "unlisted" | "local" | "private" | "direct";
// ...
    };
    responses: never;
    parameters: {
        /**
         * @description Account ID or nickname
         * @example 123
         */
        accountIdOrNickname: string;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}